types problems changed form the frontend almost
