+++ frontend/api/instructor.ts
@@ -1,9 +1,11 @@
 import { api } from "@/api/api";
-import { ApiResponse } from "@/types/apiResponse";
+import { ApiResponse } from "@/types/general";
 import {
 	IInstructorWithUserDetails,
 	InstructorFormData,
 	IInstructorDetails,
+	InstructorProfile,
+	InstructorCourse,
 } from "@/types/instructor";
 import { User } from "@/types/user";
 
@@ -262,4 +264,29 @@ export const getInstructorDetails = async (
 				error.response?.data?.message || "Failed to get instructor details",
 		};
 	}
+};
+
+
+
+
+
+export const getInstructorProfile = async (userId: string) => {
+  const response = await api.get<InstructorProfile>(`/instructors/${userId}`);
+  return response.data;
+};
+
+export const updateInstructorProfile = async (
+  userId: string,
+  data: Partial<InstructorProfile>
+) => {
+  const response = await api.patch<InstructorProfile>(
+    `/instructors/${userId}`,
+    data
+  );
+  return response.data;
+};
+
+export const getInstructorCourses = async () => {
+  const response = await api.get<InstructorCourse[]>("/instructors/courses");
+  return response.data;
 };
