--- frontend/types/apiResponse.ts
@@ -1,15 +1,0 @@
-export interface ApiResponse<T = any> {
-	statusCode: number;
-	success: boolean;
-	message: string;
-	data: T;
-	error?: string;
-}
-
-export interface IPaginatedResponse<T> {
-	items: T[];
-	total: number;
-	page: number;
-	limit: number;
-	totalPages: number;
-}
