+++ frontend/types/dashboard.ts
@@ -58,9 +58,4 @@ export interface DashboardResponse {
 	topInstructors: TopInstructor[];
 }
 
-// API Response wrapper
-export interface ApiResponse<T> {
-	success: boolean;
-	message: string;
-	data: T;
-}
+
