+++ frontend/types/general.ts
@@ -1,0 +1,24 @@
+export interface ApiResponse<T> {
+  statusCode: number;
+  success: boolean;
+  message: string;
+  data: T;
+  error?: string;
+}
+
+
+export interface ApiResponse<T = any> {
+  statusCode: number;
+  success: boolean;
+  message: string;
+  data: T;
+  error?: string;
+}
+
+export interface IPaginatedResponse<T> {
+  items: T[];
+  total: number;
+  page: number;
+  limit: number;
+  totalPages: number;
+}
