+++ frontend/types/instructor.ts
@@ -84,3 +84,95 @@ export interface IInstructorWithUserDetails {
 	updatedAt: string;
 	user: IUserDetails;
 }
+
+
+
+export interface InstructorProfile {
+  id: string;
+  userId: string;
+  firstName: string;
+  lastName: string;
+  email: string;
+  bio?: string;
+  profilePicture?: string;
+  expertise: string[];
+  totalStudents: number;
+  totalCourses: number;
+  totalEarnings: number;
+  rating: number;
+  createdAt: string;
+  updatedAt: string;
+}
+
+export interface InstructorCourse {
+  id: string;
+  title: string;
+  description: string;
+  thumbnail: string;
+  price: number;
+  students: number;
+  rating: number;
+  status: "draft" | "published" | "archived";
+  createdAt: string;
+  updatedAt: string;
+}
+
+export interface InstructorEarning {
+  id: string;
+  amount: number;
+  courseId: string;
+  courseTitle: string;
+  studentId: string;
+  studentName: string;
+  status: "pending" | "completed" | "failed";
+  createdAt: string;
+}
+
+// Instructor Dashboard Types
+export interface InstructorDashboardStats {
+  totalCourses: number;
+  totalStudents: number;
+  totalEnrollments: number;
+  totalRevenue: number;
+  activeCourses: number;
+  pendingCourses: number;
+  completedCourses: number;
+  averageRating: number;
+  totalReviews: number;
+}
+
+export interface InstructorCourseStats {
+  courseId: string;
+  courseTitle: string;
+  enrollmentCount: number;
+  revenue: number;
+  rating: number;
+  reviewCount: number;
+  status: string;
+  createdAt: string;
+  lastEnrollmentDate?: string;
+}
+
+export interface InstructorStudentStats {
+  studentId: string;
+  studentName: string;
+  email: string;
+  enrolledCourses: number;
+  lastEnrollmentDate: string;
+  isActive: boolean;
+}
+
+export interface RecentEnrollment {
+  courseId: string;
+  courseTitle: string;
+  studentName: string;
+  enrolledAt: string;
+}
+
+export interface InstructorDashboardResponse {
+  stats: InstructorDashboardStats;
+  topCourses: InstructorCourseStats[];
+  recentStudents: InstructorStudentStats[];
+  recentEnrollments: RecentEnrollment[];
+}
+
